Murders The Devil from The Bible in the form of a bee (#2672)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->
 
## About The Pull Request
Removes chem bees from the jungle and beach spawn pools, replacing them
with normal bees.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Walking off of a ship onto one of the relatively "peaceful" planets,
only to be loaded up with admin chemicals was a pretty awful experience.

We inherited these little guys from Voidcrew, honestly that should be
enough reason to remove them.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
balance: planetary chem bees have been replaced with the garden variety
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->