Vox Sprite Update (#2668)
## About The Pull Request

Updates a few Vox sprites, mostly backpacks. Adds a proper icon for the
tail hold action.

## Why It's Good For The Game

We love vox in this house

## Changelog

:cl:
imageadd: Added vox sprites & a tailhold icon.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->