Poster Printer (#2557)
## About The Pull Request

Adds poster printers. Currently unmapped, so there's no way to get them
at the moment, and I recommend one gets added to the outpost.
HOW TO USE:
get a sheet of paper and a printer

![image](https://github.com/shiptest-ss13/Shiptest/assets/64121399/44bf348e-3971-4bba-8f90-8166ca893ece)
put the paper in the printer

![image](https://github.com/shiptest-ss13/Shiptest/assets/64121399/7b157f7d-35f0-4d7b-a3aa-1a66c734c75b)
click on the printer and select the kind of poster you want to print

![image](https://github.com/shiptest-ss13/Shiptest/assets/64121399/ecab5df3-9536-464b-a0db-b3dc4ae49320)
print the poster, and then remove it

![image](https://github.com/shiptest-ss13/Shiptest/assets/64121399/b59950ab-a68a-4fd2-921b-f0baeeb0f0f0)
tada! free* poster!

![image](https://github.com/shiptest-ss13/Shiptest/assets/64121399/ba917741-2b3a-494a-923f-bee6aec7ea5a)

*actually uses some toner with each use, as well as a sheet of paper

## Why It's Good For The Game

A lot of people were complaining about how there's no easy way to get
posters at the moment, so I made a convenient printing device.

## Changelog

🆑
add: Added a poster printer. Please map this later, I didn't want to
mess with mapping for now so
/🆑

---------

Signed-off-by: Moon <64121399+MoonTheBird@users.noreply.github.com>
Signed-off-by: meem <75212565+meemofcourse@users.noreply.github.com>
Co-authored-by: meem <75212565+meemofcourse@users.noreply.github.com>