Adds a few items to the ranger. (#2675)
## About The Pull Request
Adds two flashes, a Recharger, a stamp and a wideband to the Ranger
bridge. And a commander pistol,mags and ammo to the lieutenant closet
and security closet,and adds a 9mm ammo disk for the autolathe. Also
removes the ORM due to latency's request.


## Why It's Good For The Game
It lets the captain recharge their e-rifle, complete the cyborg shell,
stamp paperwork and actually talk to people for once.
The commanders are to give a bit more firepower,as before the security
specialist had only one derringer (!!) and i feel like the lieutenant
could do with a bit more firepower,too.
The ORM was removed due to having refineries in outposts.
## Changelog

:cl:
add: Added two flashes, a recharger, a stamp, two commanders, ammo for
them and a wideband to the Ranger bridge.
del: The Ranger's ORM.
/:cl: