Valor-Class Field Care Crusier (#2144)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

![valor3](https://cdn.discordapp.com/attachments/1021926309611442246/1197668934795665499/valor.png?ex=65bc1b1d&is=65a9a61d&hm=71fadf813dda96fae87b6842e5bb8961bfd014a0f4e631513ca46e7d3e20c13d&)
![2023-07-07 21 37
25](https://cdn.discordapp.com/attachments/1021926309611442246/1197667812886458448/image.png?ex=65bc1a12&is=65a9a512&hm=988070177c49e8515f033320c8850b8d7f28dff1e10d799e569f3cf39ae3c3f8&)


Did something go terribly wrong? Do you have wounded? Are _you_ wounded?
Do you have money? You're in luck! The Valor-Class is Inteq's medical
component of their fleet, and provides a great service for a meager*
fee! Armed with a tiny patient recovery vessel and spacious medical
facilities, this flying field hospital is ready to provide rapid
tactical care for all that ails you (beyond the mental scarring of
losing all your friends). This comes at the expense of relying on other
ships for food, research or armaments.

It is staffed by:
* 1 Vanguard
* 1 Honorable Corpsman
* 3 Corpsmen
* 2 Enforcers
* 2 Recruits

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

it's the inteq medical ship it uses the inteq content

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Valor-Class Field Care Crusier
add: Haste-Class Ambulance
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: meem <75212565+meemofcourse@users.noreply.github.com>