Crashed Pubby Tweaks (#2589)
## About The Pull Request
As per the request of @Latentish 

This PR tweaks the recently merged (#2363) Crashed Pubby Ruin. The shape
of the crash area should now be more realistic. Several issues have been
fixed, as well.


![image](https://github.com/shiptest-ss13/Shiptest/assets/118859017/4dc076fe-8dce-4ab5-a37e-c201ee5e77d2)

![9cbdcebcc92d3e5d3a96e44677e129bf](https://github.com/shiptest-ss13/Shiptest/assets/118859017/9b6d7ad7-bf15-44a1-9008-4ce6ce08a3bc)

![image](https://github.com/shiptest-ss13/Shiptest/assets/118859017/d5732f90-f242-41f4-b5e7-e58b10dc5e08)


## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

🆑
tweak: The Crashed Pubby Ruin has been tweaked slightly.
fix: Fixed a floating wallmount in the Crashed Pubby Ruin.
/🆑