Paracelsus Tweaks 2 (#2640)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Reworks the layout and supply of the Paracelsus kitchen to be more
user-friendly and to better feed a wider range of patients, or just to
give the patients more to do by cooking. It now has a rolling pin, sink,
extra fridge and some meat and vegetables to cook with. The canteen
table now has a deck of cards for recreation.

![image](https://github.com/shiptest-ss13/Shiptest/assets/95449138/de121835-16db-4a90-9f84-e5a02d07c275)

Also adds two extra shark plushies because people requested them and
they help with patient recovery speeds (not confirmed).
Adds ship viewscreens to the canteen, cargo bay and scribe waiting area.
People want to see where they're going and it gives people more to do.
Adds two bookshelves to the paracelsus dorm room.
Moves the autolathe to the ghettochem area due to overwhelming popular
demand. It is necessary for ghetto chemistry and more accessible to
people as doctors are more likely to use it than an engineer or the
captain.
Fixes the locker full of guns on the bridge being just. Fully unlocked!
For some reason! It now can only be accessed by the overseer and
captain.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Paracelsus kitchen was a 7 tile nightmare for cooking. Now it is much
more user friendly and provides players with more ingredients to make
nice food with.
Doctors not having to ask to use the autolathe is also a lot better
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Reworked the paracelsus kitchen layout to be more user friendly and
have more ingredients to cook with.
add: Adds two bookcases and two extra shark plushies to the dorms.
add: Adds ship viewscreens to the paracelsus
add: Moves the autolathe on the paracelsus to the ghettochem room
fix: The firearms locker on the paracelsus no longer lacks a lock to
prevent patients from accessing the firearms
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->