Bunch of Statpanel updates from TG (#2581)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Ports a bunch of statpanel-related commits from TG. This will hopefully
reduce the amount of GetFlatIcon calls done by alt+clicking nearby
tiles, as well as just be a smoother experience overall with it. No real
functional change beyond slowing down the stat panel update time.

## Why It's Good For The Game

Smoother experience with an important part of the game window

## Changelog

:cl: Couls, Kylerace, Magatsuchi, LemonInTheDark, Fikou, Mothblocks,
Stylemistake, Mark Suckerberg
admin: Reduces MC tab update time, this can be adjusted with a
preference in OOC settings
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: Mark Suckerberg <mark@suckerberg.gay>
Co-authored-by: Couls <coul422@gmail.com>
Co-authored-by: Kylerace <kylerlumpkin1@gmail.com>
Co-authored-by: magatsuchi <88991542+magatsuchi@users.noreply.github.com>
Co-authored-by: Aleksej Komarov <stylemistake@gmail.com>
Co-authored-by: LemonInTheDark <58055496+LemonInTheDark@users.noreply.github.com>
Co-authored-by: san7890 <the@san7890.com>
Co-authored-by: Fikou <23585223+Fikou@users.noreply.github.com>
Co-authored-by: Mothblocks <35135081+Mothblocks@users.noreply.github.com>