Ship faction display take 2 (#2680)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Allows specifying a custom faction name in the ship JSON, as well as
makes it so the faction name is only inferred from prefix once per round
instead of every time the names are moused over.

Also makes ships sort into categories based on faction instead of just
being lumped into the shiptest category, so it's a bit easier to find
the ship you want to spawn.

## Why It's Good For The Game

Maptainer request

## Changelog

:cl:
admin: Ships should now be categorised by faction in the shuttle
manipulator
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->