Aegis QoL changes (community requests) (#2709)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
This is my first PR, in, like, 2 years. So maybe we're back?
EDIT: So. Initially, this was just a handful of community requests to
the Aegis.
Now it has turned into a basic SUNS-ification PR for the vessel,
stripping out all the old generic syndicate gear and replacing it all
with SUNS stuff. Also slightly nerfs the vessel by removing a few
traitor items, such as the dart pistol, night vision goggles, etc etc.
But the drip compensates.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
The ship already has SUNS loadouts, now this just translates it to
actual ship gear.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
:add: Strips out all the generic syndicate gear from the Aegis-Class and
replaces it with SUNS gear. Hooray for student debt
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->