Disables every single CMM ship (#2719)
## About The Pull Request

Disables every single CMM ship. While I wanted to do this when i was
done with the CLIP resprite, after there was a particularly bad round
involving the CMM I want to do this now.

This also deletes the corvus because fuck the corvus


@shiptest-ss13/maptainers please merge ASAP, preferably before next
uptime
## Why It's Good For The Game

It's no secret I have not been happy with the CMM players at all, they
all suck, and it hurts every day when i see a Corvus acting like indie+.
It also hurts when i can't even spawn in as a inspector or higher up
because that requires the players to roleplay, they do NOT roleplaay
which hurts a lot especially for a faction i helped create.

I have a grander plan which i've been working on, but this NEEDS to be
done

Part 0 of the CMM refresh

## Changelog

:cl:
tweak: all CMM ships are aspawn now
del: Corvus-Class
/:cl: