Adds the Cobra-20 as a buyable gun in the cargo market. (and fixes the stechkin's description) (#2703)
## About The Pull Request

Title says all, i think.
Secretly also fixes the description on the stechkin to say that it's
from scaraborough arms and not ''gorlex marauders''
## Why It's Good For The Game
Adds an option for buying an SMG in the market, since there's no SMGs
currently available to buy, which felt like a weird skip since there's
rifles, PDWs, and pistols. meant to be a sort of inbetween for from
pistols and pdws/rifles
## Changelog

:cl:
add: Adds the Cobra-20 as buyable from the outpost alongside it's
magazines.
spellcheck: fixes the stechkin being from the gorlex marauders and
returns it to it's rightful home of scaraborough arms
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->