Rebalances prices of guns. (#2756)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Rebalances the prices of guns aimed towards mostly making automatics
such as the SKM, P16 and Cobra more expensive, and slightly lowers the
price of shotguns.
Waiting until the SKM merge so that i can change the price of that
without repercussions.

## Why It's Good For The Game

Automatics are very powerful and should have a price that matches so.
Might also make non-automatic options such as the Illestren, Shotguns
and Pistols more worthwhile.

## Changelog

:cl:

balance: Cargo gun crates are more granularized, having only one gun at
a time and a price reflecting that. Prices were rebalanced making
automatics more expensive overall.

/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->