Mining Rework Part 1: First of his Name (#2446)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
In a vain attempt to make mining less boring and actually require more
player engagement I have cobbled together the groundwork to a sizable
reworking of the mining system from that thing you send out that one
random crew member to go do for a couple minutes to a team-based
activity focused around the use of specialized equipment.

The gameplay loop for mining I envision within my mind's eye is one
where most members of the ship's crew, even down to the engineers and
medics, have to cooperate in order to properly secure large amounts of
ore from dangerous environments, with most other non-mining ships
needing to at least partially rely on mining ships to supply them with
minerals.

This PR specifically adds; ore veins (they use the same sprites as
geysers pending a better one) that spawn on lava, sand, rock, and ice
planets currently and at varying levels of difficulty and yeild, a drill
with which to mine said veins, a scanner to find the aforementioned
veins, and an outpost mission that is essentially just a mission to go
drill ore veins but with money as the reward instead of ore. Oh and I
also cut surface mineral spawn rates roughly in half pending a better
gauge of the balance involved there.

Here's the harrowing story of one such man who tried to use this drill
with no armaments and no crew:

![screenshot1](https://github.com/shiptest-ss13/Shiptest/assets/112462947/1a737ff6-5372-46ab-bda3-72363dbfe94d)

![screenshot2](https://github.com/shiptest-ss13/Shiptest/assets/112462947/fb9452b8-c6cd-4e9e-8986-59bbb272bdf3)

![screenshot3](https://github.com/shiptest-ss13/Shiptest/assets/112462947/812c4c8f-d3f7-4bde-a119-e8e6f35790e8)


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Mining in its current state is just kinda bad. It's boring to actually
focus on doing and way too easy to just drown your ship in minerals.
This PR would at bare minimum partially fix that. That's not to say it's
perfect and doesn't need to be balanced and polished to hell and back,
but it's a good first step I feel.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Ore veins and a drill to mine them with
add: Mining based missions
tweak: Lowered the spawn rate of some other mob spawners like tendrils
tweak: Added the ability for mob spawners to spawn mobs a distance from
their source
balance: Cut most ore sources in half and tweaked the spawn weights
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: BogCreature <112462947+BogCreature@users.noreply.github.com>
Co-authored-by: thgvr <81882910+thgvr@users.noreply.github.com>
Co-authored-by: Mark Suckerberg <mark@suckerberg.gay>