Adds a lava canyon as a new lava planet ruin (#2642)
## About The Pull Request

Adds a lava canyon ruin, with the remains of a scorched ashwalker
village nestled between the crags. Comes with an associated area for
this ruin.

![2024-01-29 16 50
44](https://github.com/shiptest-ss13/Shiptest/assets/116136632/efe9ab52-7908-4b6f-b966-cdcae7d9d470)
![2024-01-29 16 50
41](https://github.com/shiptest-ss13/Shiptest/assets/116136632/36831083-e337-470b-8fb9-35de5d8b84cd)

A blood drunk miner defending a vault of loot!
![Screenshot 2024-01-29
165828](https://github.com/shiptest-ss13/Shiptest/assets/116136632/63214e38-316f-4dd4-a120-b418c123b7a5)


## Why It's Good For The Game

Lava planets need ruins.

## Changelog

:cl:
add: New lava canyon ruin
fix: Gate icon updates not working
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->